Height is now double for conversions from cm to inch vice versa

+++ app/src/main/java/com/mk/familyweighttracker/DbModels/UserReadingModel.java
@@ -25,7 +25,7 @@ public class UserReadingModel extends Model {
     public double Weight;
 
     @Column(name = "Height")
-    public int Height;
+    public double Height;
 
     @Column(name = "TakenOn")
     public Date TakenOn;
